backend/requirements.txt
Flask>=2.2.2
Flask-CORS>=3.0.10
Click>=8.1.3
PyYAML>=6.0

